Move tx building/spending to spend.rs, refactoring
Remove unused sp_addresses vector

Refactor create_transaction

Refactor finalize_transaction

Refactor sign_transaction
