+++ src/client/mod.rs
@@ -1,4 +1,5 @@
 mod client;
+mod spend;
 mod structs;
 
 pub use client::SpClient;
